backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
yt-dlp==2024.8.6
openai-whisper==20231117
openai==1.51.0
chromadb==0.5.5
python-dotenv==1.0.1
pydantic==2.9.2
